+++ spec/foobara/command/concerns/domain_mappers_spec.rb
@@ -8,31 +8,45 @@ RSpec.describe Foobara::Command::Concerns::DomainMappers do
       foobara_domain!
     end
   end
+
   let(:command_class) do
     rtype = result_type
     subcommand_class
+    mappers = domain_mappers
+
     stub_class("SomeDomain::SomeCommand", Foobara::Command) do
-      depends_on SomeDomain::SomeSubcommand
+      depends_on SomeDomain::SomeSubcommand, *mappers
 
       result rtype
 
       def execute
-        run_mapped_subcommand!(SomeDomain::SomeSubcommand, result_type, "bar")
+        run_mapped_subcommand!(SomeDomain::SomeSubcommand, "bar", result_type)
       end
     end
   end
+  let(:domain_mappers) do
+    [
+      SomeDomain::DomainMappers::SomeDomainMapper
+    ]
+  end
+
   let(:result_type) { :associative_array }
 
   let(:subcommand_class) do
     domain
+    include_result = include_result_type
     stub_class("SomeDomain::SomeSubcommand", Foobara::Command) do
       inputs foo: :string
+      if include_result
+        result foo: :string
+      end
 
       def execute
         inputs
       end
     end
   end
+  let(:include_result_type) { true }
 
   let(:domain_mapper) do
     subcommand_class
@@ -70,8 +84,10 @@ RSpec.describe Foobara::Command::Concerns::DomainMappers do
       let(:command_class) do
         rtype = result_type
         subcommand_class
+        mappers = domain_mappers
+
         stub_class("SomeDomain::SomeCommand", Foobara::Command) do
-          depends_on SomeDomain::SomeSubcommand
+          depends_on SomeDomain::SomeSubcommand, *mappers
 
           result rtype
 
@@ -85,11 +101,104 @@ RSpec.describe Foobara::Command::Concerns::DomainMappers do
         expect(outcome).to be_success
         expect(result).to eq(foo: "bar")
       end
+
+      context "when forgetting to depend on the mapper" do
+        let(:domain_mappers) { [] }
+
+        it "raises" do
+          expect {
+            command.run
+          }.to raise_error(Foobara::Command::Concerns::DomainMappers::ForgotToDependOnDomainMapperError)
+        end
+      end
     end
 
     context "when there's a result mapper" do
       before do
         result_domain_mapper
+      end
+
+      let(:domain_mappers) do
+        [
+          SomeDomain::DomainMappers::SomeDomainMapper,
+          SomeDomain::DomainMappers::SomeResultDomainMapper
+        ]
+      end
+
+      let(:result_domain_mapper) do
+        subcommand_class
+        domain_mapper
+        stub_module("SomeDomain::DomainMappers")
+        stub_class("SomeDomain::DomainMappers::SomeResultDomainMapper", Foobara::DomainMapper) do
+          possible_error :some_error, context: { foo: :string }, message: "kaboom!"
+
+          from SomeDomain::DomainMappers::SomeDomainMapper.to_type
+          to :string
+
+          def map
+            JSON.generate(from)
+          end
+        end
+      end
+
+      let(:result_type) { :string }
+
+      it "maps the inputs and the results" do
+        mapper = domain_mapper.domain.lookup_matching_domain_mapper! to: :string
+        expect(mapper).to be(SomeDomain::DomainMappers::SomeResultDomainMapper)
+        expect(outcome).to be_success
+        expect(result).to eq("{\"foo\":\"bar\"}")
+      end
+
+      it "contains the mapper's errors in the depending command's possible errors" do
+        possible_error_keys = command_class.possible_errors.map(&:key).map(&:to_s)
+        expect(possible_error_keys).to include(
+          "some_domain::domain_mappers::some_result_domain_mapper>runtime.some_error"
+        )
+      end
+
+      context "when the subcommand has no result type" do
+        let(:include_result_type) { false }
+
+        it "maps the inputs and the results" do
+          expect(outcome).to be_success
+          expect(result).to eq("{\"foo\":\"bar\"}")
+        end
+      end
+
+      context "when forgetting to depend on the mapper" do
+        let(:domain_mappers) { [] }
+
+        it "raises" do
+          result_domain_mapper
+          expect {
+            command.run
+          }.to raise_error(Foobara::Command::Concerns::DomainMappers::ForgotToDependOnDomainMapperError)
+        end
+      end
+    end
+
+    context "when only there's a result mapper" do
+      let(:command_class) do
+        rtype = result_type
+        subcommand_class
+        mappers = domain_mappers
+
+        stub_class("SomeDomain::SomeCommand", Foobara::Command) do
+          depends_on SomeDomain::SomeSubcommand, *mappers
+
+          result rtype
+
+          def execute
+            run_mapped_subcommand!(SomeDomain::SomeSubcommand, { foo: "bar" }, result_type)
+          end
+        end
+      end
+
+      let(:domain_mappers) do
+        [
+          result_domain_mapper
+        ]
       end
 
       let(:result_domain_mapper) do
@@ -109,8 +218,48 @@ RSpec.describe Foobara::Command::Concerns::DomainMappers do
       let(:result_type) { :string }
 
       it "maps the inputs and the results" do
+        mapper = command_class.foobara_domain.lookup_matching_domain_mapper! to: :string
+        expect(mapper).to be(SomeDomain::DomainMappers::SomeResultDomainMapper)
+        expect(command.domain_map!("bar")).to eq(foo: "bar")
         expect(outcome).to be_success
         expect(result).to eq("{\"foo\":\"bar\"}")
+      end
+
+      context "when forgetting to depend on the mapper" do
+        let(:domain_mappers) { [] }
+
+        it "raises" do
+          result_domain_mapper
+          expect {
+            command.run
+          }.to raise_error(Foobara::Command::Concerns::DomainMappers::ForgotToDependOnDomainMapperError)
+        end
+      end
+
+      context "when no mapper found" do
+        let(:domain_mappers) { [] }
+
+        let(:command_class) do
+          rtype = result_type
+          subcommand_class
+          mappers = domain_mappers
+
+          stub_class("SomeDomain::SomeCommand", Foobara::Command) do
+            depends_on SomeDomain::SomeSubcommand, *mappers
+
+            result rtype
+
+            def execute
+              run_mapped_subcommand!(SomeDomain::SomeSubcommand, { foo: "bar" }, :integer)
+            end
+          end
+        end
+
+        it "raises" do
+          expect {
+            command.run
+          }.to raise_error(Foobara::Command::Concerns::DomainMappers::NoDomainMapperFoundError)
+        end
       end
     end
   end